+++ src/constants.rs
@@ -4,7 +4,10 @@ pub const MONOSHOT_URL: &str = "https://api.moonshot.cn/v1/chat/completions";
 pub const QWEN_URL: &str = "https://dashscope.aliyuncs.com/compatible-mode/v1/chat/completions";
 pub const GPT_URL: &str = "https://api.openai.com/v1/chat/completions";
 pub const GEMINI_URL: &str = "https://generativelanguage.googleapis.com/v1beta/openai/";
-pub const DOUBAO_URL: &str = "https://ark.cn-beijing.volces.com/api/v3/chat/completions"; 
+pub const DOUBAO_URL: &str = "https://ark.cn-beijing.volces.com/api/v3/chat/completions";
+pub const GLM_URL: &str = "https://open.bigmodel.cn/api/paas/v4/chat/completions";
+pub const DEEPSEEK_URL: &str = "https://api.deepseek.com/chat/completions";
+pub const QIANFAN_URL: &str = "https://qianfan.baidubce.com/v2/chat/completions";
 
 pub const DIFF_PROMPT: &str = r#"
     You are an expert developer specialist in creating git commits.